*** account_journal_visibility ***: when we open the journals list, people will only see journal for which they are allowed (means their group is specified on the journal definition). and also Only people in the group defined on the journal will be able to see the invoices of this journal.
*** mrp_production_report ***:
 Add partner on procurement form which set when Sale order confirm , Add link between production and procument , New report for production.

*** stock_back_order ***: 
To manage all back-orders (means partial pickings):  When products coming from suppliers arrive but some are missing, we have to make a partial picking. The remaining products are called "back-orders" and have to be separated from normal waiting picking (in a predefined list called "Back-Orders"). The same process has to be done for sending goods.