[ADD] pxgo_cash_statement: New module for easier cash management.
Module for easier cash management.

Adds a "Entries by Cash Statement" view of the bank statements that shows only 
the fields required for cash management.
It works with "Cash Statement Line Types" that allow to predefine cash lines 
(pre-set the account, type [withdrawal/deposit], and description of the cash 
statement line), so the user doesn't have to cope with those details 
(even more, if the user introduces an invalid quantity, for example a positive 
amount for a withdrawal, OpenERP will automatically correct the entry).

Thanks to Igalia (http://www.igalia.com/) for sponsoring this module :)